12 working

+++ model.js
@@ -42,5 +42,13 @@ exports.updateArticles = (article_id, body) => {
 }
 
 exports.removeComment = (comment_id) => {
-    return db.query(`DELETE FROM comments WHERE comment_id=$1`, [comment_id])
+    return db.query(`SELECT * FROM comments WHERE comment_id=$1`, [comment_id])
+    .then(({rows}) => {
+        if (rows.length === 0){
+            console.log("0")
+            return Promise.reject({status: 404, message: "no comment_id found"})
+        } else {
+            db.query(`DELETE FROM comments WHERE comment_id=$1`, [comment_id])
+        }
+    })
 }
